+++ src/shader/vertex.rs
@@ -31,7 +31,7 @@ pub fn quadrilateral(top_left: Vec2, top_right: Vec2, bottom_left: Vec2, bottom_
     }
 }
 
-pub fn line(start: Vec2, end: Vec2, width: f64) -> RenderItem {
+pub fn line(start: Vec2, end: Vec2, width: f32) -> RenderItem {
     let axis = (end - start).normed().orthog();
     let bottom_left = start - axis * width / 2;
     let bottom_right = start + axis * width / 2;
@@ -46,15 +46,15 @@ pub fn line(start: Vec2, end: Vec2, width: f64) -> RenderItem {
     }
 }
 
-pub fn circle(centre: Vec2, radius: f64, steps: u32) -> RenderItem {
-    let dt = 2. * f64::PI() / f64::from(steps);
+pub fn circle(centre: Vec2, radius: f32, steps: u32) -> RenderItem {
+    let dt = 2. * f32::PI() / steps as f32;
     RenderItem {
         vertices: (0..steps).circular_tuple_windows()
             .flat_map(|(i, j)| {
                 vec![
                     centre,
-                    centre + Vec2 { x: radius * (f64::from(i) * dt).cos(), y: radius * (f64::from(i) * dt).sin() },
-                    centre + Vec2 { x: radius * (f64::from(j) * dt).cos(), y: radius * (f64::from(j) * dt).sin() },
+                    centre + Vec2 { x: radius * (i as f32 * dt).cos(), y: radius * (i as f32 * dt).sin() },
+                    centre + Vec2 { x: radius * (j as f32 * dt).cos(), y: radius * (j as f32 * dt).sin() },
                 ]
             })
             .map(VertexWithUV::from_vertex).collect(),
